T#782: Extract 6 trace routes to src/trace/routes.ts
Phase 2.4 of Library #102. Mechanical move, no logic changes.

Moved: GET /api/traces, /:id, /:id/chain, /:id/linked-chain;
POST /api/traces/:prevId/link; DELETE /api/traces/:id/link.

Imports listTraces, getTrace, getTraceChain, linkTraces, unlinkTraces,
getTraceLinkedChain from ./handler.ts (already exported).

Server.ts: -86 lines. Cumulative T#771-T#782: -6,511 lines / -54%.

T#781 was blocked on prework (30 missing exports from server.ts); filed
as T#792 by Zaghnal. T#782 trace was cleaner surface, shipped 10min.
